+++ api-builder-plugin-fn-file/src/index.js
@@ -6,8 +6,8 @@ const actions = require('./csvFileActions.js');
  * Resolves the API Builder plugin.
  * @returns {object} An API Builder plugin.
  */
-async function getPlugin() {
-	const sdk = new SDK();
+async function getPlugin(pluginConfig, options) {
+	const sdk = new SDK({ pluginConfig });
 	sdk.load(path.resolve(__dirname, 'flow-nodes.yml'), actions);
 	return sdk.getPlugin();
 }