+++ api-builder-plugin-fn-file/Changelog.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+- Added support to write native files to the disk
+- Upgrade to @axway/api-builder-sdk v1.0.0
+- Test suite now using @axway/api-builder-test-utils v1.0.0
 
 ## [0.0.2] 2020-03-30
 ### Security